refactor: Use a comprehensible word
