+++ src/global-viewer.d.ts
