+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "exclude": ["broker", "dist", "tests"],
+  "exclude": ["dist", "tests"],
   "compilerOptions": {
     "skipLibCheck": true,
     /* Basic Options */
